requirements.txt
pymongo
playwright
asyncio
motor
boto3
aiohttp
python-dotenv
pytz